refactor: extract components and set visibility to internal
This commit refactors the UI components by extracting `DaysHeaderRow` and `CalendarEventsGrid` from `SwipeableCalendarView.kt` into their own files. Additionally, the visibility of several composable functions has been changed from `public` to `internal` to encapsulate them within the module.

**Key Changes:**

-   **Component Extraction:**
    -   `DaysHeaderRow`: The composable function for displaying the day headers in calendar views has been moved from `SwipeableCalendarView.kt` to its own file, `components/DaysHeaderRow.kt`.
    -   `CalendarEventsGrid`: The composable for rendering the grid of events has been extracted from `SwipeableCalendarView.kt` and placed in a new file, `components/CalendarEventsGrid.kt`.

-   **Visibility Modifier Updates:**
    -   The following composable functions have been marked as `internal` to restrict their usage to within the same module:
        -   `SwipeablePager`
        -   `CalendarDrawer` and its helper `CalendarViewOption`
        -   `EventTag`
        -   `TimeColumn`
        -   `AddEventDialog`
        -   `CalendarTopAppBar`
        -   `BaseCalendarScreen`
        -   `SwipeableCalendarView`
        -   The newly extracted `DaysHeaderRow` and `CalendarEventsGrid`.

This refactoring improves the project's structure, making the codebase more organized and maintainable by promoting better encapsulation of UI components.
