+++ composeApp/src/commonMain/kotlin/com/debanshu/xcalendar/ui/components/SwipeableCalendarView.kt
@@ -75,7 +75,7 @@ import kotlin.time.ExperimentalTime
  * @param dynamicHeaderHeightState The height of the header row
  */
 @Composable
-fun SwipeableCalendarView(
+internal fun SwipeableCalendarView(
     modifier: Modifier = Modifier,
     startDate: () -> LocalDate,
     events: () -> List<Event>,
@@ -175,353 +175,6 @@ private fun CalendarContent(
             onEventClick = onEventClick,
             currentDate = currentDate,
             scrollState = scrollState,
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-private fun DaysHeaderRow(
-    startDate: LocalDate,
-    numDays: Int,
-    currentDate: LocalDate,
-    holidaysByDate: Map<LocalDate, List<Holiday>>,
-    onDayClick: (LocalDate) -> Unit,
-    modifier: Modifier = Modifier,
-    dynamicHeaderHeightState: MutableState<Int>?,
-) {
-    val dates =
-        List(numDays) { index ->
-            startDate.plus(DatePeriod(days = index))
-        }
-
-    Row(
-        modifier =
-            modifier
-                .background(XCalendarTheme.colorScheme.surfaceContainerLow)
-                .height(IntrinsicSize.Min)
-                .heightIn(min = 60.dp)
-                .onGloballyPositioned {
-                    if (dynamicHeaderHeightState != null) {
-                        dynamicHeaderHeightState.value = it.size.height
-                    }
-                },
-    ) {
-        if (numDays > 1) {
-            dates.forEach { date ->
-                val isToday = date == currentDate
-                val currentDayHolidays = holidaysByDate[date] ?: emptyList()
-
-                Column(
-                    modifier =
-                        Modifier
-                            .fillMaxHeight()
-                            .weight(1f)
-                            .padding(top = 8.dp)
-                            .clickable { onDayClick(date) },
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    Text(
-                        text = date.dayOfWeek.name.take(3),
-                        style = XCalendarTheme.typography.labelSmall,
-                    )
-                    Box(
-                        modifier =
-                            Modifier
-                                .padding(vertical = XCalendarTheme.dimensions.spacing_4)
-                                .clip(MaterialShapes.Cookie9Sided.toShape())
-                                .size(30.dp)
-                                .background(
-                                    when {
-                                        isToday -> XCalendarTheme.colorScheme.primary
-                                        else -> Color.Transparent
-                                    },
-                                ),
-                        contentAlignment = Alignment.Center,
-                    ) {
-                        Text(
-                            text = date.day.toString(),
-                            style = XCalendarTheme.typography.bodyMedium,
-                            color =
-                                when {
-                                    isToday -> XCalendarTheme.colorScheme.inverseOnSurface
-                                    else -> XCalendarTheme.colorScheme.onSurface
-                                },
-                        )
-                    }
-                    if (currentDayHolidays.isNotEmpty()) {
-                        Column {
-                            currentDayHolidays.take(2).forEach { holiday ->
-                                EventTag(
-                                    modifier =
-                                        Modifier
-                                            .padding(start = 4.dp, end = 4.dp, bottom = 6.dp)
-                                            .fillMaxWidth(),
-                                    text = holiday.name,
-                                    color = Color(0xFF007F73),
-                                    textColor = XCalendarTheme.colorScheme.inverseOnSurface,
-                                )
-                            }
-
-                            if (currentDayHolidays.size > 2) {
-                                val extraCount = currentDayHolidays.size - 2
-                                Text(
-                                    text = "+$extraCount more",
-                                    style = XCalendarTheme.typography.labelSmall.copy(fontSize = 8.sp),
-                                    textAlign = TextAlign.Start,
-                                    maxLines = 1,
-                                    overflow = TextOverflow.Ellipsis,
-                                    color = XCalendarTheme.colorScheme.onSurfaceVariant,
-                                    modifier =
-                                        Modifier
-                                            .padding(start = 4.dp, end = 4.dp, bottom = 6.dp)
-                                            .fillMaxWidth(),
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            val currentDayHolidays = holidaysByDate[dates.first()] ?: emptyList()
-            var holidaysExpanded by remember { mutableStateOf(false) }
-            Column(
-                modifier =
-                    Modifier
-                        .fillMaxHeight()
-                        .weight(1f),
-            ) {
-                if (currentDayHolidays.isNotEmpty()) {
-                    val displayHolidays =
-                        if (holidaysExpanded) {
-                            currentDayHolidays
-                        } else {
-                            currentDayHolidays.take(2)
-                        }
-                    displayHolidays.forEach { holiday ->
-                        Text(
-                            text = holiday.name,
-                            style = XCalendarTheme.typography.labelMedium,
-                            textAlign = TextAlign.Start,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            color = XCalendarTheme.colorScheme.inverseOnSurface,
-                            modifier =
-                                Modifier
-                                    .padding(horizontal = 4.dp, vertical = 2.dp)
-                                    .fillMaxWidth()
-                                    .background(
-                                        Color(0xFF007F73),
-                                        RoundedCornerShape(8.dp),
-                                    ).padding(8.dp),
-                        )
-                    }
-
-                    if (currentDayHolidays.size > 2 && !holidaysExpanded) {
-                        val extraCount = currentDayHolidays.size - 2
-                        Text(
-                            text = "+$extraCount more",
-                            style = XCalendarTheme.typography.labelMedium,
-                            textAlign = TextAlign.Start,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            color = XCalendarTheme.colorScheme.onSurfaceVariant,
-                            modifier =
-                                Modifier
-                                    .padding(horizontal = 8.dp, vertical = 4.dp)
-                                    .fillMaxWidth()
-                                    .clickable { holidaysExpanded = true },
-                        )
-                    } else if (holidaysExpanded && currentDayHolidays.size > 2) {
-                        Text(
-                            text = "Show less",
-                            style = XCalendarTheme.typography.labelMedium,
-                            textAlign = TextAlign.Start,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            color = XCalendarTheme.colorScheme.onSurfaceVariant,
-                            modifier =
-                                Modifier
-                                    .padding(horizontal = 8.dp, vertical = 4.dp)
-                                    .fillMaxWidth()
-                                    .clickable { holidaysExpanded = false },
-                        )
-                    }
-                    Spacer(modifier = Modifier.height(4.dp))
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalTime::class)
-@Composable
-private fun CalendarEventsGrid(
-    startDate: LocalDate,
-    numDays: Int,
-    eventsByDate: Map<LocalDate, List<Event>>,
-    timeRange: IntRange,
-    hourHeightDp: Float,
-    onEventClick: (Event) -> Unit,
-    currentDate: LocalDate,
-    scrollState: ScrollState,
-) {
-    val dates =
-        List(numDays) { index ->
-            startDate.plus(DatePeriod(days = index))
-        }
-
-    BoxWithConstraints(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .verticalScroll(scrollState)
-                .background(XCalendarTheme.colorScheme.surfaceContainerLow),
-    ) {
-        val dayColumnWidth = maxWidth / numDays
-        val now = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault())
-        val currentMinute = now.hour * 60 + now.minute
-
-        Column {
-            timeRange.forEach { _ ->
-                Row(
-                    Modifier
-                        .fillMaxWidth()
-                        .height(hourHeightDp.dp),
-                ) {
-                    repeat(numDays) {
-                        Box(
-                            Modifier
-                                .weight(1f)
-                                .border(
-                                    width = 2.dp,
-                                    color = XCalendarTheme.colorScheme.surfaceContainerLow,
-                                    shape = RoundedCornerShape(10.dp),
-                                ).fillMaxHeight()
-                                .clip(RoundedCornerShape(10.dp))
-                                .background(XCalendarTheme.colorScheme.surfaceContainerHigh),
-                        )
-                    }
-                }
-            }
-        }
-
-        if (dates.any { it == currentDate }) {
-            val dayIndex = dates.indexOfFirst { it == currentDate }
-            if (dayIndex >= 0) {
-                val offsetX = dayColumnWidth * dayIndex
-                val offsetY = (currentMinute / 60f * hourHeightDp).dp
-
-                Box(
-                    modifier =
-                        Modifier
-                            .offset(x = offsetX, y = offsetY)
-                            .width(dayColumnWidth)
-                            .height(2.dp)
-                            .background(XCalendarTheme.colorScheme.primary),
-                )
-            }
-        }
-
-        // Process events by date using pre-calculated mapping
-        dates.forEachIndexed { dayIndex, date ->
-            val dayEvents = eventsByDate[date] ?: emptyList()
-
-            // Group overlapping events with caching
-            val eventGroups = remember(dayEvents) { groupOverlappingEvents(dayEvents) }
-
-            eventGroups.forEach { (_, group) ->
-                val totalOverlapping = group.size
-
-                group.forEachIndexed { _, event ->
-                    val eventStart =
-                        event.startTime.toLocalDateTime(TimeZone.currentSystemDefault())
-                    val eventEnd = event.endTime.toLocalDateTime(TimeZone.currentSystemDefault())
-                    val hour = eventStart.hour
-                    val minute = eventStart.minute
-
-                    if (hour in timeRange) {
-                        val durationMinutes =
-                            if (eventStart.date == eventEnd.date) {
-                                (eventEnd.hour - hour) * 60 + (eventEnd.minute - minute)
-                            } else {
-                                (24 - hour) * 60 - minute
-                            }
-
-                        val topOffset =
-                            (hour - timeRange.first) * hourHeightDp + (minute / 60f) * hourHeightDp
-                        val eventHeight = (durationMinutes / 60f) * hourHeightDp
-
-                        // Calculate width and horizontal position based on overlap
-                        EventItem(
-                            event = event,
-                            onClick = { onEventClick(event) },
-                            modifier =
-                                Modifier
-                                    .offset(
-                                        x = dayColumnWidth * dayIndex,
-                                        y = topOffset.dp,
-                                    ).width(dayColumnWidth)
-                                    .height(eventHeight.dp.coerceAtLeast(30.dp))
-                                    .padding(1.dp),
-                            isOverlapping = totalOverlapping > 1,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun groupOverlappingEvents(events: List<Event>): Map<Int, List<Event>> {
-    val sortedEvents = events.sortedBy { it.startTime }
-    val groups = mutableMapOf<Int, MutableList<Event>>()
-    var groupId = 0
-
-    sortedEvents.forEach { event ->
-        val eventStart = event.startTime
-        val eventEnd = event.endTime
-
-        val existingGroup =
-            groups.entries.firstOrNull { (_, groupEvents) ->
-                groupEvents.none {
-                    (eventStart < it.endTime && eventEnd > it.startTime)
-                }
-            }
-
-        if (existingGroup != null) {
-            existingGroup.value.add(event)
-        } else {
-            groups[groupId] = mutableListOf(event)
-            groupId++
-        }
-    }
-
-    return groups
-}
-
-@Composable
-private fun EventItem(
-    event: Event,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    isOverlapping: Boolean = false,
-) {
-    Box(
-        modifier =
-            modifier
-                .clip(RoundedCornerShape(4.dp))
-                .border(1.dp, color = Color(event.color))
-                .background(Color(event.color).copy(alpha = if (isOverlapping) 0.7f else 0.9f))
-                .clickable(onClick = onClick)
-                .padding(4.dp),
-    ) {
-        Text(
-            text = event.title,
-            style = XCalendarTheme.typography.labelSmall,
-            color = XCalendarTheme.colorScheme.inverseOnSurface,
-            maxLines = 2,
-            overflow = TextOverflow.Ellipsis,
         )
     }
 }
